Add an optimised approach to double PLT indirection.
This uses a global variable in lib.c to track determine if
the internal_message has been called in this library.  If
it has, then it directly calls the implementation from that
point on.  Otherwise, it will continue to call internal_message.

This means that the common case will be a test of a global and a call
that could even be inlined.